requirements.txt
streamlit 
selenium 
json 
time 
google-generativeai
